Share metadata state key helper
